+++ abci/did/v1/dataStruct.go
@@ -225,9 +225,10 @@ type GetAsNodesByServiceIdWithNameResult struct {
 }
 
 type InitNDIDParam struct {
-	NodeID          string `json:"node_id"`
-	PublicKey       string `json:"public_key"`
-	MasterPublicKey string `json:"master_public_key"`
+	NodeID           string `json:"node_id"`
+	PublicKey        string `json:"public_key"`
+	MasterPublicKey  string `json:"master_public_key"`
+	ChainHistoryInfo string `json:"chain_history_info"`
 }
 
 type TransferNDIDParam struct {